Changes location of the files
