--- app/(pages)/(auth)/signIn/page.tsx
@@ -1,14 +1,0 @@
-import SignIn from './SignIn'
-
-export const metadata = {
-  title: 'Evently SignIn',
-  description: 'Evently signIn to your account',
-}
-
-export default function SignInPage() {
-  return (
-    <div className="m-10">
-      <SignIn />
-    </div>
-  )
-}
