--- app/@modal/(.)signIn/page.tsx
@@ -1,4 +1,0 @@
-import SignIn from '@/app/(pages)/(auth)/signIn/SignIn'
-export default function SignInModal() {
-  return <SignIn />
-}
