+++ src/app/(pages)/(auth)/signIn/page.tsx
@@ -1,0 +1,5 @@
+import SignInForm from '@/ui/auth/SignInForm'
+
+export default function SignInPage() {
+  return <SignInForm />
+}
