+++ src/app/@modal/(.)signIn/page.tsx
@@ -1,0 +1,10 @@
+import SignInForm from '@/ui/auth/SignInForm'
+
+export const metadata = {
+  title: 'Evently SignIn',
+  description: 'Evently signIn to your account',
+}
+
+export default function SignInModal() {
+  return <SignInForm />
+}
