+++ src/app/@modal/layout.tsx
@@ -1,0 +1,6 @@
+import React, { ReactNode } from 'react'
+import Modal from '@/ui/Modal/Modal'
+
+export default function LayoutModal({ children }: { children: ReactNode }) {
+  return <Modal>{children}</Modal>
+}
