+++ src/ui/Header/SearchBar/SearchBar.tsx
@@ -1,0 +1,38 @@
+'use client'
+import { WhiteMagnifier } from '@/ui/Icons'
+import SearchByWordsInput from './SearchInput'
+import PlacesAutoCompleteInput from './PlacesAutoCompleteInput'
+
+interface SearchBarProps {
+  mobile?: boolean
+}
+
+export default function SearchBar({ mobile = false }: SearchBarProps) {
+  // PENDING IMPLEMENTATION
+  // const searchByWords = async (formData: FormData) => {
+  //   'use server'
+  // }
+
+  const isMobile = mobile ? 'mt-10 mx-auto lg:hidden' : 'hidden'
+
+  return (
+    <form
+      //action= PENDING IMPLMENTATION
+      className={`realtive grow w-full sm:max-w-xl flex sm:flex-row items-center justify-center lg:flex lg:flex-row lg:items-center ${isMobile}`}
+      onSubmit={(event) => event.preventDefault()}
+    >
+      <div className="sm:flex w-full">
+        <SearchByWordsInput />
+        <PlacesAutoCompleteInput />
+      </div>
+
+      <button
+        type="submit"
+        className="self-stretch  bg-red-500 rounded-r-lg border border-red-400 hover:bg-red-500 focus:z-10focus:border-red-400 flex-col items-center justify-center p-3"
+        aria-label="Search events"
+      >
+        <WhiteMagnifier />
+      </button>
+    </form>
+  )
+}
