+++ src/ui/Icons/Avatar/Avatar.tsx
