+++ src/ui/Icons/GreatMagnifier/GreyMagnifier.tsx
