+++ src/ui/Icons/WhiteManifier/WhiteManifier.tsx
